MirroringTable: refactor, move method, extract WriteOperationInfo
